publicrustbot-main/requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
requests==2.31.0
flask==3.0.2
PyNaCl==1.5.0
yt-dlp==2024.3.10 